ADDED: test to add inch and cm

+++ src/main/java/com/bl/QuantityMeasurement/model/Centimeters.java
@@ -1,7 +1,7 @@
 package com.bl.QuantityMeasurement.model;
 
 public class Centimeters extends LengthUnit {
-    public Centimeters(int length) {
+    public Centimeters(double length) {
         super(length, LengthUnit.UnitType.CM);
     }
 }
